Refactors Renderes to render module
