+++ lib/devlogs/render/log_template_renderer.rb
@@ -1,0 +1,8 @@
+require_relative 'erb_template_renderer'
+
+#
+# LogTemplateRenderer captures log information and 
+# renders it within a given ERB template 
+#
+class LogTemplateRenderer < ErbTemplateRenderer 
+end
